Rename: #21 layout 구분을 위하여 page 디렉터리 추가 및 관련 파일 이동
